+++ projects/ng-flowchart/src/lib/model/flow.model.ts
@@ -1,4 +1,3 @@
-import { EmbeddedViewRef } from '@angular/core';
 import { NgFlowchartCanvasService } from '../ng-flowchart-canvas.service';
 import { NgFlowchartStepComponent } from '../ng-flowchart-step/ng-flowchart-step.component';
 
@@ -40,85 +39,14 @@ export namespace NgFlowchart {
          * Clears all flow chart, reseting the current canvas
          */
         clear() {
-            // if (this.canvas.canvasData?.rootElement) {
-            //     this.canvas.canvasData.rootElement.destroy(true, false);
-            //     this.canvas.reRender();
-            // }
+            if (this.canvas.flow?.rootStep) {
+                this.canvas.flow.rootStep.destroy(true, false);
+                this.canvas.reRender();
+            }
 
         }
 
-    }
-
-    // export class Step {
-    //     constructor(private canvas: NgFlowchartCanvasService, private step: NgFlowchartAbstractStep) {
-
-    //     }
-
-    //     toJSON() {
-    //         return this.step.toJSON();
-    //     }
-
-    //     /**
-    //      * Returns the id of this step. This is the same as the html id of the element.
-    //      * To make modifications to the view you can getElementById
-    //      */
-    //     get id() {
-    //         return this.step.id;
-    //     }
-
-    //     /**
-    //      * Returns the referenced data object passed via the ngFlowchartStepData input
-    //      */
-    //     get data() {
-    //         return this.step.data;
-    //     }
-
-    //     /**
-    //      * Returns the parent step. This can be null if this is the root node.
-    //      */
-    //     getParent() {
-    //         return new Step(this.canvas, this.step.getParent());
-    //     }
-
-    //     hasChildren() {
-    //         return this.step.hasChildren();
-    //     }
-
-    //     isRootStep() {
-    //         return !this.getParent();
-    //     }
-
-    //     /**
-    //      * Returns all direct children of this step. 
-    //      * To get all descendants you can recursively keep calling getChildren
-    //      */
-    //     getChildren(): Array<Step> {
-    //         return this.step.getChildren().map(child => {
-    //             return new Step(this.canvas, child)
-    //         })
-    //     }
-
-    //     /**
-    //      * Deletes this node from the tree. Returns true or false if delete was a success
-    //      * @param recursive Should its children also be deleted? (default is false)
-    //      */
-    //     delete(recursive: boolean = false) {
-    //         // let result = this.canvasElement.destroy(recursive);
-    //         // if (result) {
-    //         //     this.canvas.reRender();
-    //         // }
-    //         // return result;
-    //     }
-
-        
-        
-    // }
-
-    
-
-    export interface StepView extends EmbeddedViewRef<any> {
-        data?: any
-    }
+    }    
 
     export class Options {
         /** The gap (in pixels) between flow steps*/
