+++ projects/ng-flowchart/src/lib/ng-flowchart-step/ng-flowchart-step.component.html
@@ -1,7 +1,14 @@
-<div #canvasContent>
+<div #canvasContent [id]="id">
   <ng-container
-    *ngTemplateOutlet="contentTemplate; context: { $implicit: data, id: id }"
+    *ngTemplateOutlet="
+      contentTemplate;
+      context: {
+        $implicit: {
+          data: data,
+          id: id
+        }
+      }
+    "
   >
   </ng-container>
-  
 </div>
