+++ projects/ng-flowchart/src/lib/services/dropdata.service.ts
@@ -1,10 +1,9 @@
 import { Injectable, TemplateRef, Type } from "@angular/core";
-import { NgFlowchartAbstractStep } from '../ng-flowchart-step/ng-flowchart-abstract-step';
 import { NgFlowchartStepComponent } from '../ng-flowchart-step/ng-flowchart-step.component';
 
 export type DragStep = {
     template?: TemplateRef<any> | Type<NgFlowchartStepComponent>,
-    instance?: NgFlowchartAbstractStep,
+    instance?: NgFlowchartStepComponent,
     data?: any
 }
