+++ projects/workspace/debug.log
@@ -1,0 +1,3 @@
+[0103/150348.781:ERROR:crash_report_database_win.cc(496)] failed to stat report
+[0103/151100.606:ERROR:crash_report_database_win.cc(496)] failed to stat report
+[0103/151138.694:ERROR:crash_report_database_win.cc(496)] failed to stat report
