+++ projects/workspace/src/app/app.component.ts
@@ -85,6 +85,6 @@ export class AppComponent {
   }
 
   onDelete(id) {
-    this.canvasElement.getFlow().getStep(id).destroy(true);
+    this.canvasElement.getFlow().getStep(id).destroy(false);
   }
 }
